add "username" field in TicketEntity, refactor dao package
