+++ src/test/java/ClientDAOTest.java
@@ -1,12 +1,9 @@
-import DAO.ClientEntity;
-import DAO.ClientRepository;
-import DAO.FlightEntity;
-import DAO.FlightRepository;
+import dao.entities.ClientEntity;
+import dao.repositories.ClientRepository;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.json.simple.parser.ParseException;
 import org.junit.jupiter.api.Test;
-import service.Client;
 
 import java.io.File;
 import java.io.IOException;
