+++ src/test/java/FlightDAOTest.java
@@ -1,5 +1,5 @@
-import DAO.FlightEntity;
-import DAO.FlightRepository;
+import dao.entities.FlightEntity;
+import dao.repositories.FlightRepository;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.json.simple.parser.ParseException;
